Reimplement duplicate column handling while awake
I was tired and in a hurry when I implemented `--replace`, and I didn't
like the code. Here's a new version with a more flexible
`--duplicate-columns` flag and much cleaner code.

This also adds an integration test suite which actually runs the tool
from the command line and uses the SmartyStreets API.